feat: :sparkles: Shared calendar component for desktop and mobile
